+++ libraries/session/src/main/java/androidx/media3/session/DefaultMediaNotificationProvider.java
@@ -401,9 +401,10 @@ public class DefaultMediaNotificationProvider implements MediaNotification.Provi
    * <p>Override this method to customize the buttons on the notification. Commands of the buttons
    * returned by this method must be contained in {@link MediaController#getAvailableCommands()}.
    *
-   * <p>By default, the notification shows {@link Player#COMMAND_PLAY_PAUSE} in {@linkplain
-   * Notification.MediaStyle#setShowActionsInCompactView(int...) compact view}. This can be
-   * customized by defining the index of the command in compact view of up to 3 commands in their
+   * <p>By default, the notification shows buttons for {@link Player#seekToPreviousMediaItem()},
+   * {@link Player#play()} or {@link Player#pause()}, {@link Player#seekToNextMediaItem()} in
+   * {@linkplain Notification.MediaStyle#setShowActionsInCompactView(int...) compact view}. This can
+   * be customized by defining the index of the command in compact view of up to 3 commands in their
    * extras with key {@link DefaultMediaNotificationProvider#COMMAND_KEY_COMPACT_VIEW_INDEX}.
    *
    * <p>To make the custom layout and commands work, you need to {@linkplain
@@ -491,9 +492,11 @@ public class DefaultMediaNotificationProvider implements MediaNotification.Provi
    * and define which actions are shown in compact view by returning the indices of the buttons to
    * be shown in compact view.
    *
-   * <p>By default, {@link Player#COMMAND_PLAY_PAUSE} is shown in compact view, unless some of the
-   * buttons are marked with {@link DefaultMediaNotificationProvider#COMMAND_KEY_COMPACT_VIEW_INDEX}
-   * to declare the index in compact view of the given command button in the button extras.
+   * <p>By default, the buttons for {@link Player#seekToPreviousMediaItem()}, {@link Player#play()}
+   * or {@link Player#pause()}, {@link Player#seekToNextMediaItem()} are shown in compact view,
+   * unless some of the buttons are marked with {@link
+   * DefaultMediaNotificationProvider#COMMAND_KEY_COMPACT_VIEW_INDEX} to declare the index in
+   * compact view of the given command button in the button extras.
    *
    * @param mediaSession The media session to which the actions will be sent.
    * @param mediaButtons The command buttons to be included in the notification.
@@ -509,7 +512,9 @@ public class DefaultMediaNotificationProvider implements MediaNotification.Provi
       NotificationCompat.Builder builder,
       MediaNotification.ActionFactory actionFactory) {
     int[] compactViewIndices = new int[3];
+    int[] defaultCompactViewIndices = new int[3];
     Arrays.fill(compactViewIndices, INDEX_UNSET);
+    Arrays.fill(defaultCompactViewIndices, INDEX_UNSET);
     int compactViewCommandCount = 0;
     for (int i = 0; i < mediaButtons.size(); i++) {
       CommandButton commandButton = mediaButtons.get(i);
@@ -534,10 +539,26 @@ public class DefaultMediaNotificationProvider implements MediaNotification.Provi
       if (compactViewIndex >= 0 && compactViewIndex < compactViewIndices.length) {
         compactViewCommandCount++;
         compactViewIndices[compactViewIndex] = i;
-      } else if (commandButton.playerCommand == COMMAND_PLAY_PAUSE
-          && compactViewCommandCount == 0) {
-        // If there is no custom configuration we use the play/pause action in compact view.
-        compactViewIndices[0] = i;
+      } else if (commandButton.playerCommand == COMMAND_SEEK_TO_PREVIOUS
+          || commandButton.playerCommand == COMMAND_SEEK_TO_PREVIOUS_MEDIA_ITEM) {
+        defaultCompactViewIndices[0] = i;
+      } else if (commandButton.playerCommand == COMMAND_PLAY_PAUSE) {
+        defaultCompactViewIndices[1] = i;
+      } else if (commandButton.playerCommand == COMMAND_SEEK_TO_NEXT
+          || commandButton.playerCommand == COMMAND_SEEK_TO_NEXT_MEDIA_ITEM) {
+        defaultCompactViewIndices[2] = i;
+      }
+    }
+    if (compactViewCommandCount == 0) {
+      // If there is no custom configuration we use the seekPrev (if any), play/pause (if any),
+      // seekNext (if any) action in compact view.
+      int indexInCompactViewIndices = 0;
+      for (int i = 0; i < defaultCompactViewIndices.length; i++) {
+        if (defaultCompactViewIndices[i] == INDEX_UNSET) {
+          continue;
+        }
+        compactViewIndices[indexInCompactViewIndices] = defaultCompactViewIndices[i];
+        indexInCompactViewIndices++;
       }
     }
     for (int i = 0; i < compactViewIndices.length; i++) {